[add]add handle_files separated interfacing files
